requirements.txt
streamlit
openai
pymongo
python-dotenv
sentence-transformers
pandas
PyPDF2
scikit-learn
numpy